# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for personal art shop
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Artwork {
  id: number;
  title: string;
  description: string;
  category: string;
  price: number;
  imageUrl: string;
  available: boolean;
  year: number;
  dimensions: string;
  medium: string;
  featured: boolean;
  createdAt: string;
}

export interface CreateArtworkBody {
  title: string;
  description: string;
  category: string;
  price: number;
  imageUrl: string;
  available: boolean;
  year: number;
  dimensions: string;
  medium: string;
  featured: boolean;
}

export interface ContactMessageBody {
  name: string;
  email: string;
  subject: string;
  message: string;
}

export type ListArtworksParams = {
  category?: string;
  available?: boolean;
};

export type SendContactMessage200 = {
  success: boolean;
};