# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for personal art shop
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all artworks
 */
export const ListArtworksQueryParams = zod.object({
  category: zod.coerce.string().optional(),
  available: zod.coerce.boolean().optional(),
});

export const ListArtworksResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  category: zod.string(),
  price: zod.number(),
  imageUrl: zod.string(),
  available: zod.boolean(),
  year: zod.number(),
  dimensions: zod.string(),
  medium: zod.string(),
  featured: zod.boolean(),
  createdAt: zod.coerce.date(),
});
export const ListArtworksResponse = zod.array(ListArtworksResponseItem);

/**
 * @summary Create a new artwork
 */
export const CreateArtworkBody = zod.object({
  title: zod.string(),
  description: zod.string(),
  category: zod.string(),
  price: zod.number(),
  imageUrl: zod.string(),
  available: zod.boolean(),
  year: zod.number(),
  dimensions: zod.string(),
  medium: zod.string(),
  featured: zod.boolean(),
});

/**
 * @summary Get featured artworks for homepage
 */
export const GetFeaturedArtworksResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  category: zod.string(),
  price: zod.number(),
  imageUrl: zod.string(),
  available: zod.boolean(),
  year: zod.number(),
  dimensions: zod.string(),
  medium: zod.string(),
  featured: zod.boolean(),
  createdAt: zod.coerce.date(),
});
export const GetFeaturedArtworksResponse = zod.array(
  GetFeaturedArtworksResponseItem,
);

/**
 * @summary List all unique categories
 */
export const ListCategoriesResponseItem = zod.string();
export const ListCategoriesResponse = zod.array(ListCategoriesResponseItem);

/**
 * @summary Get a single artwork by ID
 */
export const GetArtworkParams = zod.object({
  id: zod.coerce.number(),
});

export const GetArtworkResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  category: zod.string(),
  price: zod.number(),
  imageUrl: zod.string(),
  available: zod.boolean(),
  year: zod.number(),
  dimensions: zod.string(),
  medium: zod.string(),
  featured: zod.boolean(),
  createdAt: zod.coerce.date(),
});

/**
 * @summary Send a contact message
 */
export const SendContactMessageBody = zod.object({
  name: zod.string(),
  email: zod.string(),
  subject: zod.string(),
  message: zod.string(),
});

export const SendContactMessageResponse = zod.object({
  success: zod.boolean(),
});