# lib/api-zod/src/generated/types/artwork.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification for personal art shop
 * OpenAPI spec version: 0.1.0
 */

export interface Artwork {
  id: number;
  title: string;
  description: string;
  category: string;
  price: number;
  imageUrl: string;
  available: boolean;
  year: number;
  dimensions: string;
  medium: string;
  featured: boolean;
  createdAt: Date;
}